Refactor ID token validation out into `githubactions` package
