+++ lib/utils/githubactions/githubactions.go
@@ -1,15 +1,5 @@
 package githubactions
 
-import (
-	"context"
-	"encoding/json"
-	"github.com/gravitational/trace"
-	"io"
-	"net/http"
-	"net/url"
-	"os"
-)
-
 // GitHub Workload Identity
 //
 // GH provides workloads with two environment variables to faciliate fetching
@@ -31,81 +21,6 @@ import (
 // - https://docs.github.com/en/actions/deployment/security-hardening-your-deployments/configuring-openid-connect-in-cloud-providers
 
 const IssuerURL = "https://token.actions.githubusercontent.com"
-
-type tokenResponse struct {
-	Value string `json:"value"`
-}
-
-// IDTokenSource allows a GitHub ID token to be fetched whilst executing
-// within the context of a GitHub actions workflow.
-type IDTokenSource struct {
-	getIDTokenURL   func() string
-	getRequestToken func() string
-	client          http.Client
-}
-
-func NewIDTokenSource() *IDTokenSource {
-	return &IDTokenSource{
-		getIDTokenURL: func() string {
-			return os.Getenv("ACTIONS_ID_TOKEN_REQUEST_URL")
-		},
-		getRequestToken: func() string {
-			return os.Getenv("ACTIONS_ID_TOKEN_REQUEST_TOKEN")
-		},
-	}
-}
-
-// GetIDToken utilises values set in the environment and the GitHub API to
-// fetch a GitHub issued IDToken.
-func (ip *IDTokenSource) GetIDToken(ctx context.Context) (string, error) {
-	// TODO: Inject audience to be set cluster name instead
-	audience := "teleport.cluster.local"
-
-	tokenURL := ip.getIDTokenURL()
-	requestToken := ip.getRequestToken()
-	if tokenURL == "" {
-		return "", trace.BadParameter(
-			"ACTIONS_ID_TOKEN_REQUEST_URL environment variable missing",
-		)
-	}
-	if requestToken == "" {
-		return "", trace.BadParameter(
-			"ACTIONS_ID_TOKEN_REQUEST_TOKEN environment variable missing",
-		)
-	}
-
-	tokenURL = tokenURL + "&audience=" + url.QueryEscape(audience)
-	req, err := http.NewRequestWithContext(
-		ctx, http.MethodGet, tokenURL, nil,
-	)
-	if err != nil {
-		return "", trace.Wrap(err)
-	}
-	req.Header.Set("Authorization", "Bearer "+requestToken)
-	req.Header.Set("Accept", "application/json; api-version=2.0")
-	req.Header.Set("Content-Type", "application/json")
-	res, err := ip.client.Do(req)
-	if err != nil {
-		return "", trace.Wrap(err)
-	}
-	defer res.Body.Close()
-
-	bytes, err := io.ReadAll(res.Body)
-	if err != nil {
-		return "", trace.Wrap(err)
-	}
-
-	var data tokenResponse
-	if err := json.Unmarshal(bytes, &data); err != nil {
-		return "", trace.Wrap(err)
-	}
-
-	if data.Value == "" {
-		return "", trace.Errorf("response did not include ID token")
-	}
-
-	return data.Value, nil
-}
 
 // IDTokenClaims is the structure of claims contained within a Github issued
 // ID token.
@@ -138,6 +53,7 @@ type IDTokenClaims struct {
 	// The visibility of the repository where the workflow is running. Accepts the following values: internal, private, or public.
 	RepositoryVisibility string `json:"repository_visibility"`
 	// The repository from where the workflow is running.
+	// This includes the name of the owner e.g `gravitational/teleport`
 	Repository string `json:"repository"`
 	// The ID of the repository from where the workflow is running.
 	RepositoryID string `json:"repository_id"`
